feat: optimize dashboard layout
