+++ src/pages/profile/index.tsx
@@ -102,127 +102,125 @@ export default function ProfilePage() {
   }
 
   return (
-    <div className="mx-auto max-w-7xl px-4 py-4 lg:px-12">
-      <div className="mx-auto flex max-w-7xl flex-col space-y-6 lg:flex-row lg:space-x-6 lg:space-y-0">
-        {/* 左侧布局 */}
-        <div className="w-full space-y-6 lg:w-1/3">
-          {/* 个人信息 Card */}
-          <Card className="rounded-lg shadow-lg">
-            <div className="flex items-center space-x-6 p-4">
-              <div className="h-16 w-16 flex-shrink-0 overflow-hidden rounded-full">
-                <Avatar className="h-full w-full">
-                  <AvatarImage src={profile?.avatarUrl} alt="Avatar" className="object-cover" />
-                  <AvatarFallback>{profile?.displayName?.charAt(0)}</AvatarFallback>
-                </Avatar>
-              </div>
-              <div className="flex-grow">
-                <h2 className="text-2xl font-semibold leading-relaxed text-white">
-                  {profile?.displayName || profile?.username}
-                </h2>
-                <p className="leading-relaxed text-gray-400">{profile?.bio || 'No bio...'}</p>
-                <div className="mt-2 flex space-x-4">
+    <div className="mx-auto flex max-w-7xl flex-col space-y-6 lg:flex-row lg:space-x-6 lg:space-y-0">
+      {/* 左侧布局 */}
+      <div className="w-full space-y-6 lg:w-1/3">
+        {/* 个人信息 Card */}
+        <Card className="rounded-lg shadow-lg">
+          <div className="flex items-center space-x-6 p-4">
+            <div className="h-16 w-16 flex-shrink-0 overflow-hidden rounded-full">
+              <Avatar className="h-full w-full">
+                <AvatarImage src={profile?.avatarUrl} alt="Avatar" className="object-cover" />
+                <AvatarFallback>{profile?.displayName?.charAt(0)}</AvatarFallback>
+              </Avatar>
+            </div>
+            <div className="flex-grow">
+              <h2 className="text-2xl font-semibold leading-relaxed text-white">
+                {profile?.displayName || profile?.username}
+              </h2>
+              <p className="leading-relaxed text-gray-400">{profile?.bio || 'No bio...'}</p>
+              <div className="mt-2 flex space-x-4">
+                <a
+                  href={`https://github.com/${profile?.username}`}
+                  target="_blank"
+                  rel="noopener noreferrer"
+                  className="text-gray-400 hover:text-gray-500"
+                >
+                  <Github size={24} />
+                </a>
+                {profile?.twitterUsername && (
                   <a
-                    href={`https://github.com/${profile?.username}`}
+                    href={`https://twitter.com/${profile?.twitterUsername}`}
                     target="_blank"
                     rel="noopener noreferrer"
                     className="text-gray-400 hover:text-gray-500"
                   >
-                    <Github size={24} />
+                    <Twitter size={24} />
                   </a>
-                  {profile?.twitterUsername && (
-                    <a
-                      href={`https://twitter.com/${profile?.twitterUsername}`}
-                      target="_blank"
-                      rel="noopener noreferrer"
-                      className="text-gray-400 hover:text-gray-500"
-                    >
-                      <Twitter size={24} />
-                    </a>
-                  )}
+                )}
 
-                  <a
-                    href={`https://opencampus-codex.blockscout.com/address/${linkedAddress}`}
-                    className="text-gray-400 hover:text-gray-500"
-                  >
-                    <Wallet size={24} />
-                  </a>
-                </div>
+                <a
+                  href={`https://opencampus-codex.blockscout.com/address/${linkedAddress}`}
+                  className="text-gray-400 hover:text-gray-500"
+                >
+                  <Wallet size={24} />
+                </a>
               </div>
             </div>
+          </div>
 
-            <div className="grid grid-cols-2 items-center justify-between rounded-b-lg bg-gray-900 p-4">
-              <div className="text-center">
-                <h3 className="text-2xl font-bold text-white">{profile?.followers}</h3>
-                <p className="text-gray-400">Followers</p>
-              </div>
-              <div className="text-center">
-                <h3 className="text-2xl font-bold text-white">{profile?.following}</h3>
-                <p className="text-gray-400">Following</p>
-              </div>
+          <div className="grid grid-cols-2 items-center justify-between rounded-b-lg bg-gray-900 p-4">
+            <div className="text-center">
+              <h3 className="text-2xl font-bold text-white">{profile?.followers}</h3>
+              <p className="text-gray-400">Followers</p>
             </div>
-          </Card>
+            <div className="text-center">
+              <h3 className="text-2xl font-bold text-white">{profile?.following}</h3>
+              <p className="text-gray-400">Following</p>
+            </div>
+          </div>
+        </Card>
 
-          <Card className="space-y-1 rounded-lg p-4 shadow-lg">
-            <div className="space-y-1">
-              <h3 className="text-lg font-semibold text-white">Points</h3>
-              <div className="flex items-center justify-between">
-                <p className="text-l mb-2 font-bold text-gray-400">{userPoints}</p>
-              </div>
+        <Card className="space-y-1 rounded-lg p-4 shadow-lg">
+          <div className="space-y-1">
+            <h3 className="text-lg font-semibold text-white">Points</h3>
+            <div className="flex items-center justify-between">
+              <p className="text-l mb-2 font-bold text-gray-400">{userPoints}</p>
             </div>
-            <div className="space-y-1">
-              <h3 className="text-lg font-semibold text-white">Rewards</h3>
-              <div className="flex items-center justify-between">
-                <p className="text-l mb-2 font-bold text-gray-400">{totalRewards.toFixed(5)} EDU</p>
-              </div>
+          </div>
+          <div className="space-y-1">
+            <h3 className="text-lg font-semibold text-white">Rewards</h3>
+            <div className="flex items-center justify-between">
+              <p className="text-l mb-2 font-bold text-gray-400">{totalRewards.toFixed(5)} EDU</p>
             </div>
-            <div className="space-y-1">
-              <h3 className="text-lg font-semibold text-white">To Claim</h3>
-              <div className="flex items-center justify-between">
-                <p className="text-l mb-2 font-bold text-gray-400">{(totalRewards - claimedRewards).toFixed(5)} EDU</p>
-                <Button onClick={handleClaim} size="sm" disabled={claiming}>
-                  Claim{claiming && <Loader2 className="ml-2 h-4 w-4 animate-spin" />}
-                </Button>
-              </div>
+          </div>
+          <div className="space-y-1">
+            <h3 className="text-lg font-semibold text-white">To Claim</h3>
+            <div className="flex items-center justify-between">
+              <p className="text-l mb-2 font-bold text-gray-400">{(totalRewards - claimedRewards).toFixed(5)} EDU</p>
+              <Button onClick={handleClaim} size="sm" disabled={claiming}>
+                Claim{claiming && <Loader2 className="ml-2 h-4 w-4 animate-spin" />}
+              </Button>
             </div>
-            <div className="space-y-1">
-              <h3 className="text-lg font-semibold text-white">Skills</h3>
-              <div className="flex flex-wrap">
-                <span className="m-1 rounded-full bg-gray-700 px-3 py-1 text-xs text-white">C++</span>
-                <span className="m-1 rounded-full bg-gray-700 px-3 py-1 text-xs text-white">Java</span>
-                <span className="m-1 rounded-full bg-gray-700 px-3 py-1 text-xs text-white">Python</span>
-              </div>
+          </div>
+          <div className="space-y-1">
+            <h3 className="text-lg font-semibold text-white">Skills</h3>
+            <div className="flex flex-wrap">
+              <span className="m-1 rounded-full bg-gray-700 px-3 py-1 text-xs text-white">C++</span>
+              <span className="m-1 rounded-full bg-gray-700 px-3 py-1 text-xs text-white">Java</span>
+              <span className="m-1 rounded-full bg-gray-700 px-3 py-1 text-xs text-white">Python</span>
             </div>
-          </Card>
-        </div>
+          </div>
+        </Card>
+      </div>
 
-        {/* 右侧布局 */}
-        <div className="w-full space-y-6 lg:w-2/3">
-          {/* GitHub 贡献图 Card */}
-          <Card className="rounded-lg p-4 shadow-lg">
-            <h3 className="mb-2 text-lg font-semibold text-white">Contributions</h3>
-            <GitHubCalendar username={profile?.username || ''} />
-          </Card>
+      {/* 右侧布局 */}
+      <div className="w-full space-y-6 lg:w-2/3">
+        {/* GitHub 贡献图 Card */}
+        <Card className="rounded-lg p-4 shadow-lg">
+          <h3 className="mb-2 text-lg font-semibold text-white">Contributions</h3>
+          <GitHubCalendar username={profile?.username || ''} />
+        </Card>
 
-          {/* 活动流 Card */}
-          <Card className="rounded-lg p-4 shadow-lg">
-            <h3 className="mb-2 text-lg font-semibold text-white">Recent Activities</h3>
-            <ul className="space-y-4">
-              <li className="flex items-center space-x-4">
-                <span className="text-gray-400">[PR] Merged PR #42 in youbet-task</span>
-                <span className="text-xs text-gray-500">2 hours ago</span>
-              </li>
-              <li className="flex items-center space-x-4">
-                <span className="text-gray-400">[Issue] Opened Issue #7 in youbet-task</span>
-                <span className="text-xs text-gray-500">4 hours ago</span>
-              </li>
-              <li className="flex items-center space-x-4">
-                <span className="text-gray-400">[Comment] Commented on Issue #7 in youbet-task</span>
-                <span className="text-xs text-gray-500">6 hours ago</span>
-              </li>
-              {/* 继续添加更多的活动 */}
-            </ul>
-          </Card>
-        </div>
+        {/* 活动流 Card */}
+        <Card className="rounded-lg p-4 shadow-lg">
+          <h3 className="mb-2 text-lg font-semibold text-white">Recent Activities</h3>
+          <ul className="space-y-4">
+            <li className="flex items-center space-x-4">
+              <span className="text-gray-400">[PR] Merged PR #42 in youbet-task</span>
+              <span className="text-xs text-gray-500">2 hours ago</span>
+            </li>
+            <li className="flex items-center space-x-4">
+              <span className="text-gray-400">[Issue] Opened Issue #7 in youbet-task</span>
+              <span className="text-xs text-gray-500">4 hours ago</span>
+            </li>
+            <li className="flex items-center space-x-4">
+              <span className="text-gray-400">[Comment] Commented on Issue #7 in youbet-task</span>
+              <span className="text-xs text-gray-500">6 hours ago</span>
+            </li>
+            {/* 继续添加更多的活动 */}
+          </ul>
+        </Card>
       </div>
     </div>
   )
