+++ src/pages/pull-request/index.tsx
@@ -145,9 +145,5 @@ function PullRequestsTable(): React.ReactElement {
 }
 
 export default function PullRequestAdmin() {
-  return (
-    <div className="mx-auto max-w-7xl px-4 py-4 lg:px-12">
-      <PullRequestsTable />
-    </div>
-  )
+  return <PullRequestsTable />
 }
